[Refactor] Split QueryContextManager from QueryContext
