+++ be/src/agent/task_worker_pool.cpp
@@ -60,6 +60,7 @@
 #include "common/thread/thread.h"
 #include "common/util/misc.h"
 #include "exec/pipeline/query_context.h"
+#include "exec/pipeline/query_context_manager.h"
 #include "exec/workgroup/work_group.h"
 #include "fs/fs_util.h"
 #include "gen_cpp/DataCache_types.h"
