+++ be/src/exec/CMakeLists.txt
@@ -184,6 +184,7 @@ set(PIPELINE_RUNTIME_SRCS
     pipeline/exec_state_reporter.cpp
     pipeline/fragment_context.cpp
     pipeline/query_context.cpp
+    pipeline/query_context_manager.cpp
     pipeline/operator_factory.cpp
     pipeline/source_operator.cpp
     pipeline/schedule/timeout_tasks.cpp
