+++ be/src/exec/pipeline/pipeline_driver_poller.cpp
@@ -20,6 +20,7 @@
 #include "exec/pipeline/pipeline_fwd.h"
 #include "exec/pipeline/pipeline_metrics.h"
 #include "exec/pipeline/query_context.h"
+#include "exec/pipeline/query_context_manager.h"
 #include "exec/pipeline/schedule/event_scheduler.h"
 #include "runtime/current_thread.h"
 #include "runtime/exec_env.h"
