+++ be/src/runtime/profile_report_worker.cpp
@@ -17,6 +17,7 @@
 #include "common/config_exec_flow_fwd.h"
 #include "common/util/misc.h"
 #include "exec/pipeline/query_context.h"
+#include "exec/pipeline/query_context_manager.h"
 #include "runtime/fragment_mgr.h"
 
 namespace starrocks {
