+++ be/test/exec/pipeline/pipeline_file_scan_node_test.cpp
@@ -41,6 +41,7 @@
 #include "exec/pipeline/pipeline_builder.h"
 #include "exec/pipeline/pipeline_driver_executor.h"
 #include "exec/pipeline/query_context.h"
+#include "exec/pipeline/query_context_manager.h"
 #include "exec/pipeline/scan/connector_scan_operator.h"
 #include "gen_cpp/InternalService_types.h"
 #include "gtest/gtest.h"
