+++ be/test/exec/pipeline/pipeline_test_base.cpp
@@ -23,6 +23,7 @@
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/group_execution/execution_group_builder.h"
 #include "exec/pipeline/pipeline_driver_executor.h"
+#include "exec/pipeline/query_context_manager.h"
 #include "exec/workgroup/work_group.h"
 #include "exprs/function_context.h"
 #include "runtime/chunk_helper.h"
