+++ be/test/runtime/external_scan_context_mgr_test.cpp
@@ -42,6 +42,7 @@
 #include "compute_env/compute_env.h"
 #include "compute_env/result/result_queue_mgr.h"
 #include "exec/pipeline/query_context.h"
+#include "exec/pipeline/query_context_manager.h"
 #include "runtime/exec_env.h"
 #include "runtime/fragment_mgr.h"
