+++ be/test/runtime/fragment_mgr_test.cpp
@@ -46,6 +46,7 @@
 #include "common/util/thrift_util.h"
 #include "exec/data_sink.h"
 #include "exec/pipeline/query_context.h"
+#include "exec/pipeline/query_context_manager.h"
 #include "gen_cpp/Descriptors_types.h"
 #include "gen_cpp/Exprs_types.h"
 #include "gen_cpp/PlanNodes_types.h"
